refactor: reorganiza o projeto em uma estrurutura modular para uma SPA navegavel atrv´és de desafios isolados
