+++ src/challenges/Slider/Slide.js
