+++ src/challenges/Slider/SlideItem.js
