+++ src/challenges/Slider/style.js
